requirements.txt
absl-py==2.1.0
addict==2.4.0
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.8.0
astor==0.8.1
basicsr==1.4.2
beautifulsoup4==4.13.3
certifi==2024.12.14
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
Cython==3.0.12
decorator==5.1.1
deepdiff==8.4.2
eval_type_backport==0.2.2
exceptiongroup==1.2.2
facexlib==0.3.0
filelock==3.18.0
filterpy==1.4.5
fire==0.7.0
fonttools==4.56.0
fsspec==2025.3.0
future==1.0.0
gfpgan==1.3.8
grpcio==1.71.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.6.2
lxml==5.3.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.4
MouseInfo==0.1.3
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
orderly-set==5.3.0
packaging==24.2
paddleocr==2.10.0
paddlepaddle==2.6.2
pillow==11.1.0
platformdirs==4.3.6
protobuf==3.20.2
ProtoNet==0.1
psutil==7.0.0
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pydantic==2.10.6
pydantic_core==2.27.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytweening==1.2.0
PyYAML==6.0.2
RapidFuzz==3.12.2
realesrgan==0.3.0
requests==2.32.3
scikit-image==0.24.0
scipy==1.13.1
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stringzilla==3.12.3
sympy==1.13.1
tb-nightly==2.20.0a20250314
tensorboard-data-server==0.7.2
termcolor==2.5.0
tifffile==2024.8.30
tomli==2.2.1
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
yapf==0.43.0
zipp==3.21.0
